feat(pipeline): refactor context handling by moving context building logic to ctxrunner
